pyproject.toml
[project]
name = "mindora-adk-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.0",
    "google-adk>=1.9.0",
    "pandas>=2.3.1",
    "rouge-score>=0.1.2",
    "tabulate>=0.9.0",
    "uvicorn>=0.35.0",
]
